+++ lib/healthpod.dart
@@ -30,7 +30,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:solidui/solidui.dart';
 
-import 'package:healthpod/providers/settings.dart';
 import 'package:healthpod/theme/app_theme.dart';
 import 'package:healthpod/utils/create_solid_login.dart';
 
@@ -73,10 +72,6 @@ class _HealthPodState extends ConsumerState<HealthPod> {
 
   @override
   Widget build(BuildContext context) {
-    // Initialise settings.
-
-    ref.watch(settingsInitializerProvider);
-
     return SolidThemeApp(
       title: 'Solid Health Pod',
       theme: AppTheme.lightTheme,